pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["oohlearn <learnooh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
pip = "^24.2"
django-taggit = "^6.0.0"
django-tinymce = "^4.1.0"
mindpowered-reservedseating = "^0.0.5"
django-shortuuidfield = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
